requirements.txt
aioelasticsearch==0.5.2
aiohttp==3.4.4
aiomongodel==0.2.0
aioredis==1.2.0
attrs==18.2.0
tornado==4.2
aiojobs
bs4
Celery==4.4.2
cssselect==1.0.3
demjson==2.2.4
more-itertools==4.3.0
motor==2.0.0
multidict==4.5.2
parsel==1.5.1
py==1.7.0
pycryptodome==3.6.6
pyDes==2.0.1
PyExecJS==1.5.1
pymongo==3.7.2
python-dateutil==2.7.5
pytz==2018.7
pyasn1
redis
rsa==4.0
Twisted==18.9.0
w3lib==1.19.0
Werkzeug==0.14.1
zope.interface==4.6.0
pyppeteer
retrying
selenium
urllib3==1.24.1
requests==2.20.1
requests-toolbelt==0.8.0
python-redis-lock